Add hacky PrivateHyperLogLogEstimator
This is a horrifying implementation that duplicates a lot of the functionality from Airlift's DenseHll structure but with the Laplace mechanism stuffed in the middle of the cardinality estimation. Exposes a function private_cardinality(hll, epsilon) that returns the DP cardinality from a standard HLL. Currently has considerable bias issues, due at least in part to not implementing the bias correction from the Airlift implementation.